Optimized catalog to prevent unnecessary re-renders
